requirements.txt
torch
segmentation-models-pytorch
flask
pillow
numpy
